+++ addOns/spiderAjax/src/main/java/org/zaproxy/zap/extension/spiderAjax/ExtensionAjax.java
@@ -134,6 +134,8 @@ public class ExtensionAjax extends ExtensionAdaptor {
             getSpiderPanel().stopScan();
             getSpiderPanel().unload();
 
+            SpiderEventPublisher.unregisterPublisher();
+
             getView()
                     .getMainFrame()
                     .getMainFooterPanel()