+++ addOns/spiderAjax/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update minimum ZAP version to 2.9.0.
 - Maintenance changes.
 
+### Fixed
+- Unregister the event publisher when the add-on is uninstalled.
+
 ## [23.1.0] - 2020-01-17
 ### Added
 - Add repo URL.